+++ attendancedetails/src/main/java/com/employee/attendance/model/Employee.java
@@ -1,5 +1,6 @@
 package com.employee.attendance.model;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 import java.util.Date;
@@ -10,7 +11,7 @@ import java.util.Date;
 public class Employee {
 	
 	public static final String SEQUENCE_NAME = "users_sequence";
-	private long  id;
+	private BigInteger  id;
 	private long  employeeId;
 	private String firstName;
 	private String lastName;
@@ -19,11 +20,11 @@ public class Employee {
 	
 	private Date date;
 
-	public long  getId() {
+	public BigInteger  getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(BigInteger id) {
 		this.id = id;
 	}
